remove some notifications
