+++ OCKSample/AppDelegate.swift
@@ -36,91 +36,32 @@ import ParseSwift
 import UIKit
 import WatchConnectivity
 
-class AppDelegate: UIResponder, UIApplicationDelegate, ObservableObject {
+class AppDelegate: UIResponder, ObservableObject {
 
+    // MARK: Public read/write properties
+    @Published var isFirstAppOpen = true
+
+    // MARK: Public read private write properties
     let isSyncingWithCloud = true // True to sync with ParseServer, False to Sync with iOS Watch
-    var isFirstAppOpen = true
-    private var sessionDelegate: SessionDelegate!
-    private lazy var watch = OCKWatchConnectivityPeer()
-    private(set) var parseRemote: ParseRemote!
-    private(set) var store: OCKStore?
     // swiftlint:disable:next line_length
     @Published private(set) var storeManager: OCKSynchronizedStoreManager = .init(wrapping: OCKStore(name: Constants.noCareStoreName,
                                                                                                      type: .inMemory)) {
         willSet {
             StoreManagerKey.defaultValue = newValue
             DispatchQueue.main.async {
-                NotificationCenter.default.post(.init(name: Notification.Name(rawValue: Constants.storeInitialized)))
+                self.objectWillChange.send()
             }
         }
     }
+    private(set) var parseRemote: ParseRemote!
+    private(set) var store: OCKStore?
     private(set) var healthKitStore: OCKHealthKitPassthroughStore!
 
-    func application(_ application: UIApplication,
-                     didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
+    // MARK: Private read/write properties
+    private var sessionDelegate: SessionDelegate!
+    private lazy var watch = OCKWatchConnectivityPeer()
 
-        // Parse-Server setup
-        PCKUtility.setupServer(fileName: Constants.parseConfigFileName) { _, completionHandler in
-            completionHandler(.performDefaultHandling, nil)
-        }
-        return true
-    }
-
-    func application(_ application: UIApplication,
-                     configurationForConnecting connectingSceneSession: UISceneSession,
-                     options: UIScene.ConnectionOptions) -> UISceneConfiguration {
-        if isSyncingWithCloud {
-            if User.current != nil {
-                Logger.appDelegate.info("User is already signed in...")
-                guard let uuid = ProfileViewModel.getRemoteClockUUIDAfterLoginFromLocalStorage() else {
-                    Logger.appDelegate.info("Error in SceneDelage, no uuid saved.")
-                    return UISceneConfiguration(name: "Default Configuration",
-                                                sessionRole: connectingSceneSession.role)
-                }
-                setupRemotes(uuid: uuid)
-                parseRemote.automaticallySynchronizes = true
-                DispatchQueue.main.asyncAfter(deadline: .now() + 1.0) {
-                    NotificationCenter.default.post(.init(name: Notification.Name(rawValue: Constants.requestSync)))
-                }
-            }
-        } else {
-            // When syncing directly with watchOS, we do not care about login and need to setup remotes
-            setupRemotes()
-            Task {
-                do {
-                    try await store?.populateSampleData()
-                    try await healthKitStore.populateSampleData()
-                    DispatchQueue.main.asyncAfter(deadline: .now() + 1.0) {
-                        self.healthKitStore.requestHealthKitPermissionsForAllTasksInStore { error in
-
-                            if error != nil {
-                                Logger.appDelegate.error("\(error!.localizedDescription)")
-                            }
-                        }
-                    }
-                } catch {
-                    Logger.appDelegate.error("""
-                        Error in SceneDelage, could not populate
-                        data stores: \(error.localizedDescription)
-                    """)
-                }
-            }
-        }
-        return UISceneConfiguration(name: "Default Configuration", sessionRole: connectingSceneSession.role)
-    }
-
-    func application(_ application: UIApplication, didDiscardSceneSessions sceneSessions: Set<UISceneSession>) {}
-
-    func application(_ application: UIApplication,
-                     didRegisterForRemoteNotificationsWithDeviceToken deviceToken: Data) {
-        Task {
-            await Utility.updateInstallationWithDeviceToken(deviceToken)
-        }
-    }
-}
-
-// MARK: Helpers
-extension AppDelegate {
+    // MARK: Helpers
     func resetAppToInitialState() {
         NotificationCenter.default.post(.init(name: Notification.Name(rawValue: Constants.storeDeinitialized)))
         do {
